refactor: shard, check shard coverage on upload
